Resolve circular decorator import
