+++ appscale/tools/remote_helper.py
@@ -27,20 +27,7 @@ from agents.gce_agent import CredentialTypes
 from agents.gce_agent import GCEAgent
 from local_state import APPSCALE_VERSION
 from local_state import LocalState
-
-
-def threaded(func):
-  """ Decorator that allows a function to run in a separate thread.
-
-  Args:
-    func: The function to be run.
-  """
-  def _threaded(*args, **kwargs):
-    """ Executes the given function in a thread. """
-    thread = threading.Thread(target=func, args=args, kwargs=kwargs)
-    thread.start()
-    return thread
-  return _threaded
+from utils import threaded
 
 
 class RemoteHelper(object):
